Slight changes to the PID compute function arguments and fixes to convertOutputToPWM.
PID is now more generic by computing over the error instead of the
difference between reference and input. This way it can be used for
angles, speed, etc; convertOutputToPWM changed to respect the user
defined limits while still accommodating for a particular limit in a
cycle. In this commit, the user limits are set before the test loop for
the speed PID starts and the power supply voltage is read before each
compute cycle.